STY: move ci test class
