+++ pysat/tests/classes/cls_ci_tests.py
